requirements.txt
Flask
opencv-python
Pillow
easyocr
symspellpy
numpy